fix(runtime): vier stille Signaturabweichungen und "Print" ohne Argument (BUG-44)
scripts/compare_commands.py haelt commands.h gegen die Befehlsliste des
Originals (blitzcc +k: 536 Befehle, bei uns 304) und vergleicht Stelligkeit,
Rueckgabetyp und die Reihenfolge der Parameter. Der letzte Punkt ist der
wichtigste: Blitz3D kennt keine benannten Argumente, eine vertauschte
Reihenfolge ist deshalb ein stilles Falschergebnis und faellt sonst nirgends
auf. Vier davon gab es:

- JoyDown und JoyHit nahmen (port, button), das Original nimmt
  (button[,port]). Ein gueltiges Programm fragte damit den falschen Knopf ab.
  Port steht jetzt hinten und ist optional.
- ReadBytes und WriteBytes nahmen (file, bank, ...), das Original
  (bank, file, ...) - die Uebertragung lief still in die falsche Richtung.
  Beide liefern jetzt zusaetzlich die Zahl der uebertragenen Bytes.
- Floor und Ceil lieferten int, das Original meldet "Floor# ( float# )". Aus
  "Ceil(3.2) / 3" wurde dadurch eine Ganzzahldivision: 1 statt 1.33333. Der
  Kommentar im Quelltext behauptete ausdruecklich das Gegenteil ("matches
  Blitz3D's integer-output semantics") und ist richtiggestellt.

Dazu die haeufigste Ablehnung: "Print" ohne Argument meldete "Not enough
parameters", obwohl das Original "Print [string$]" fuehrt. Es gibt jetzt eine
Leerzeile aus.

Nachweis (gemessen): Leerzeile wird gedruckt; Floor(3.7)=3, Ceil(3.2)=4,
Ceil(3.2)/3=1.33333, Floor(7.9)/2=3.5; acht Bytes ueber Bank und Datei
geschrieben und zurueckgelesen, beide Werte unveraendert. Gegenprobe: derselbe
Test scheitert mit dem Compiler aus 0321de7 an "Not enough parameters for
'Print'".

Absicherung: Aequivalenzpruefung ueber alle 115 tests/*.bb und beide
examples/**/*.bb - 78 von 78 Emittaten byte-identisch, Ablehnungen 39 zu 38;
die eine weniger ist der neue Test. Diagnosevergleich ueber alle 35 neg_*.bb:
null Abweichungen. Weil eine Runtime-Aenderung im Emittat gar nicht sichtbar
wird, sind die betroffenen Programme zusaetzlich gebaut und ausgefuehrt:
test_m17_math bleibt zeichenweise gegen .expected gruen, test_m33_joystick
laeuft mit korrigierter Argumentreihenfolge. Alle 23 .expected_err gruen,
IDE-Smoke 14 von 14, gen_commands.py --check sauber.

Die Restliste steht im Buglist-Eintrag nach Klassen sortiert: 21 Befehle, bei
denen wir strenger sind als die Sprache (der Buffer-Default braucht Zustand in
der Runtime), fuenf mit zu kleinem Maximum, und WaitKey/SeekFile, die im
Original einen Wert liefern - bei WaitKey fehlt dafuer Funktionalitaet, nicht
nur eine Signatur.

+++ src/compiler/bb_math.h
@@ -38,9 +38,12 @@ inline float bb_Log(float x)   { return std::log(x); }
 inline float bb_Log10(float x) { return std::log10(x); }
 inline float bb_Exp(float x)   { return std::exp(x); }
 
-// Floor/Ceil return int — matches Blitz3D's integer-output semantics for console
-inline int   bb_Floor(float x) { return static_cast<int>(std::floor(x)); }
-inline int   bb_Ceil(float x)  { return static_cast<int>(std::ceil(x)); }
+// Floor/Ceil liefern float. Der Kommentar hier behauptete das Gegenteil
+// ("matches Blitz3D's integer-output semantics") - gemessen am Original meldet
+// `blitzcc +k` aber `Floor# ( float# )` und `Ceil# ( float# )`. Mit int als
+// Rueckgabetyp wurde aus `a# = Ceil(x#) / 2` eine Ganzzahldivision (BUG-44).
+inline float bb_Floor(float x) { return std::floor(x); }
+inline float bb_Ceil(float x)  { return std::ceil(x); }
 
 // Int(double/int) — truncate toward zero (distinct from Int(string) in bb_string.h).
 // double overload: handles both float (implicit float→double) and double literals.
